requirements.txt
flask==2.3.3
langchain==0.1.0
langchain-openai==0.0.2
langchain-experimental==0.0.47
langchain-core==0.1.7
openai==1.6.1
pandas==2.1.4
python-dotenv==1.0.0
werkzeug==2.3.7
numpy==1.24.3
qrcode==7.4.2
Pillow==10.1.0
tabulate==0.9.0